01_financialAgent_Phidata/requirements.txt
phidata
python-dotenv
yfinance
packaging
duckduckgo-search
fastapi
uvicorn 
openai
groq # groq has hosted open source libraries